Revert "Sync fork last m ain"

+++ shared/src/schemas/database/WorkoutPlanAssignmentSets.zod.ts
@@ -16,8 +16,7 @@ export const workoutPlanAssignmentSetsSchema = z.object({
   set_type: z.string().nullable(),
   reps: z.number().nullable(),
   weight: z.number().nullable(),
-  // Per-set duration is integer SECONDS.
-  duration: z.number().int().nullable(),
+  duration: z.number().nullable(),
   rest_time: z.number().nullable(),
   notes: z.string().nullable(),
   created_at: z.date().nullable(),
@@ -31,8 +30,7 @@ export const workoutPlanAssignmentSetsInitializerSchema = z.object({
   set_type: z.string().optional().nullable(),
   reps: z.number().optional().nullable(),
   weight: z.number().optional().nullable(),
-  // Per-set duration is integer SECONDS.
-  duration: z.number().int().optional().nullable(),
+  duration: z.number().optional().nullable(),
   rest_time: z.number().optional().nullable(),
   notes: z.string().optional().nullable(),
   created_at: z.date().optional().nullable(),
@@ -46,8 +44,7 @@ export const workoutPlanAssignmentSetsMutatorSchema = z.object({
   set_type: z.string().optional().nullable(),
   reps: z.number().optional().nullable(),
   weight: z.number().optional().nullable(),
-  // Per-set duration is integer SECONDS.
-  duration: z.number().int().optional().nullable(),
+  duration: z.number().optional().nullable(),
   rest_time: z.number().optional().nullable(),
   notes: z.string().optional().nullable(),
   created_at: z.date().optional().nullable(),
